Added Fritzing design, states and other changes. (Currently this wont compile)

+++ src/iTrashCan_prototype/ColorDefinitions.h
@@ -1,10 +1,15 @@
+//
+// ©Copyright 2018, All Rights Reserved.
+//
+// ColorDefinitions.h created by spinomishra on 10/29/2018T10:49 PM
+//
 #pragma once
 
 struct Color {
 	uint16_t R;
 	uint16_t G;
 	uint16_t B;
-	int WhiteLight;
+	uint16_t WhiteLight;
 };
 
 enum ColorMarkers {
